Delay actually running OMXeditor until we are on the app MainLoop.
* OMXeditor/__init__.py: we are reading a file (or asking the user to select
one), during instantiation of the OMXeditorApp and before starting the
MainLoop.  This creates a bunch of windows, and sends triggers which will only
be processed once in the MainLoop.  So we use CallAfter to delay this until
we are on MainLoop.  This prevents crashing if an error happens since it will
be caught in the MainLoop.