Drop use of wx.aui.AuiManager on the main window.
* we have a single window in the main window frame, a notebook with multiple
control panels.  The AuiManager is overkill for our requirements since it
has nothing to manager, and caused delay when resizing windows.  Dropping
in favour of simplicity.  It was not even adjusting the window sizes to
display all the controls, in windows at least it misses the last row of
buttons.