Fix positioning of viewerWindows.
* In Windows 7 (at least on micronb), the ZY viewer window was being placed
outside the monitor.  The problem was because all viewer windows are
positioned based on the original position of the mainWindow.  However, this
position was being recorded before the window was displayed which made that
viewer be positioned outside.  We fix this by undoing this crazy logic. We
leave wx to decide where to put the mainWindow (and wx will leave it up to
the window manager).  We leave creation of the XY viewer also up to wx and
move it only if we need a ZY viewer on its left.  We also no longer move the
mainWindow, the viewer windows do NOT need to be positioned relative to it.
* also drop keep the parent property of control panel as it already has a
Parent property that does the same (although we will now use the GetParent
method).
* support args, and kwargs for the control panel, and remove forced styles,
default positions, etc. Leave that shit up to the window manager.