Create toolbar from the row of buttons previously at the bottom of each page.
* The row of buttons at the bottom of each notebook page was not very
obvious. In windows 7 it was even outside the default view and not
obvious that window needed resize. We remove it and instead create a
toolbar that is managed by the main window.
* We also don't need the very same buttons repeated on each page so
having them on the toolbar makes more sense.
* like the menus, the mainwindow enable and disables the button if there
is a page open.
* we lose the help text that was displayed on the help window when
hovering over the button. But really, if we need such descriptive help
text, then we need to redesign our interface.  Anyway, these are now
docs of the methods and we can access them through __doc__ properties
(we will need to use functools.wraps() to decorate it otherwise we get
the docs of the decorated).
* Change the default of the menu_item function to require_open since
most of our menus, do require an open page.