Allow closing of the view control window (actually hides it).
* the "show view controls" menu item, actually toggled view of the the
ViewControlWindow.  The later started hidden, and had no close button on
the window.  Restore the close button (remove the non-defauls style).
Since we must need to keep the window around, we simply hide it when the
user closes it.  The menu entry now actually does what it says, i.e., show
the window and will no longer hide it if it's already open.