Fix batch processing dialog to properly display buttons.
The batch processing dialog has a fixed size which was probably good for
the system where it was designed.  At the moment, at least with Windows 7,
only the text and checkboxes are displayed.  The window needs to be resized
to see the OK, cancel buttons.  Most of the users don't know this and get
stuck.

We fixed this by creating a proper MultiChoiceDialog instead of a simple
dialog with our own statictext and checkboxes.  The huge text forced us
to use wx.lib.wordwrap.

We also drop the events ok and cancel and use ShowModal instead and we no
longer have to handle destruction of the dialog window.  Also, the logic
of interpreting the results of the dialog and preparing the crop belongs
to the mainwindow, not the dialog, so all of that gets moved out dialogs.py.

Probably none of this was required. The huge and completely inappropriate
help text is little more than noise.  An application should not require
so much information and reading for interactive usage.  And the checkboxes
are also too much.  If the point is to apply the configuration from the
current image, *all* of them should be applied, not part. This just makes
things more complicated for little gain.

Finally, we should probably have a crop/align_parameters class, which would
make things simpler. The mainwindow would simply pass such an object to
datadoc which would perform the modifications, mainwindow should not be
handling this.